Make TextComponent a container
